+++ pext/__main__.py
@@ -1740,22 +1740,14 @@ class ViewModel():
 
             return
 
-        # If longer and no new entries, only filter existing list
-        if len(self.last_search) > 0 and len(search_string) > len(self.last_search):
-            filter_entry_list = self.sorted_filtered_entry_list
-            filter_command_list = self.sorted_filtered_command_list
-        else:
-            filter_entry_list = self.sorted_entry_list
-            filter_command_list = self.sorted_command_list
-
         self.filtered_entry_list = []
         self.filtered_command_list = []
 
-        for entry in filter_entry_list:
+        for entry in self.sorted_entry_list:
             if search_string in entry.lower():
                 self.filtered_entry_list.append(entry)
 
-        for command in filter_command_list:
+        for command in self.sorted_command_list:
             if search_string in command.lower():
                 self.filtered_command_list.append(command)